Script for dependency reasons
This script is tethered to a node used in Game.tscn scene, I forgot to take it out, but it is supposed to be attached to the AlarmBell2D Sprite Node, to turn on or blare as desired from the UML document. Unfortunately, have not been able to implement it yet as it has not been showing how I want it to come out.